api/requirements.txt
# Django & API
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3

# Channels & WebSockets
channels>=4.0
channels-redis>=4.2
daphne>=4.0

# Redis
redis>=5.0

# MongoDB (match logs, optional storage)
pymongo>=4.6

# Utils
python-dotenv>=1.0
gunicorn>=21.0
